Made breadcrumb into a separate component Closes #42
