+++ src/layouts/index.js
@@ -8,7 +8,6 @@ import { Wrapper } from '../components/wrapper'
 import './index.css'
 
 const Layout = ({ children, data }) => (
-
   <div>
     <Helmet
       title={data.site.siteMetadata.title}
