+++ src/templates/diagnosis.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { Wrapper } from "../components/wrapper"
 import Link from 'gatsby-link';
-import styled from "styled-components"
+import styled from "styled-components";
+import Breadcrumb from "../components/breadcrumb";
 
 const Back = styled.span`
  position: fixed;
@@ -12,33 +13,14 @@ const Back = styled.span`
  text-decoration: underline;
  font-weight: 700;
 `
-const Margin = styled.div`
-  margin-top: 2rem;
-  margin-bottom: 1rem;
-`
+
 
 export default ({ data }) => {
   const condition = data.markdownRemark;
-  const currentURL = condition.fields.slug;
-  let breadcrumb = currentURL.split("/").filter(Boolean);
-  breadcrumb.shift();
-  breadcrumb.pop();
 
   return (
     <Wrapper>
-      <Margin>
-        <Link to="/">Home</Link>{
-          breadcrumb.map((item, i) => {
-            const url = currentURL.split(item)[0] + item;
-            return (
-              <span>
-                <span> > </span>
-                <Link key={`key-${i}`} to={url}>{item}</Link>
-              </span>
-            )
-          })
-        }
-      </Margin>
+      <Breadcrumb currentURL={condition.fields.slug} />
       <h1>{condition.frontmatter.title}</h1>
       <div dangerouslySetInnerHTML={{ __html: condition.tableOfContents }} />
